# Added Utils::split (and tests).  Added performance test for string splitting.