Based on timing results (see below), we should do two things:
	1) Switch to the HTSJDK `ParsingUtils::split` method for all
	   cases where we split a string using a single character.
	2) Switch to the `Utils::split` method for all cases where
	   we split a string by another string of length > 1.

A quick stopgap that will reduce splitting time by ~1/2 is to just
replace all calls to Java's `String::split` with `Utils::split` (since
they both take two Strings as arguments, it should be very easy).

Also, I have disabled the test so that it doesn't slow down testing cycles.

Fixes #3759

--------------------------------------------------------------------------------
Timing Results:
--------------------------------------------------------------------------------
Java Split String:
	Words Total Time:	131867865203ns	131867.865203ms	Per Split:	6048.98464233945ns		0.00604898464233945ms
	"Chars" Total Time:	12917243085ns	12917.243085ms	Per Split:	3004.010019767442ns		0.003004010019767442ms
HTSJDK ParsingUtils::split
	"Chars" Total Time:	5882790859ns	5882.790859ms	Per Split:	1368.0908974418605ns		0.0013680908974418606ms
GATK Utils::split
	Words Total Time:	38734463275ns	38734.463275ms	Per Split:	1776.8102419724771ns		0.0017768102419724772ms
	"Chars" Total Time:	7120052467ns	7120.052467ms	Per Split:	1655.826155116279ns		0.0016558261551162792ms
================================================================================